catalog: move catalog cache and methods into BTree (item 132)
Add IndexInfo, CatalogSnapshot, and all catalog lookup/mutation methods
directly to BTree. BTree now bootstraps the catalog on an empty file.
Catalog becomes a thin Deref wrapper around BTree for backward compat.
btree_graph::dump now takes &BTree directly. IndexInfo re-exported from
storage module via btree instead of catalog. Zero warnings.
