+++ src/storage.rs
@@ -13,7 +13,6 @@ mod btree;
 mod btree_graph;
 mod btree_verify;
 
-pub use crate::catalog::IndexInfo;
 pub use btree::decode_integer_key;
 pub use btree::decode_u64_key;
 pub use btree::encode_index_value;
@@ -21,6 +20,7 @@ pub use btree::encode_integer_key;
 pub use btree::encode_u64_key;
 pub use btree::BTree;
 pub use btree::CursorHandle;
+pub use btree::IndexInfo;
 pub use cell_reader::CellReader;
 pub use node_page_store::NodePageStore;
 pub use page_id::{Error as StorageError, PageId};
